MADS: Add Macintosh resource loader
Note that this is completely untested. The Macintosh version does not
work at all. The resources appear to have the same size as their PC
equivalents (for the most part), but I haven't compared the binary data
so there may be endian issues.

The index loader maps the Mac resources to the same names as the DOS
resources, but there are a couple that I don't know what they are, and
others that aren't accounted for. The cursor resources could probably be
read from the Mac executable, but the font resources appear to be

MESSAGES.DAT appears to be split into multiple Mnnn resources, so they
aren't handled either.

Sounds are stored in Rex Sound Data, but there is no attempt to read
that one either. Judging by the names of those resources they are
digitized sound effects. I have no idea how the music is handled.

I really don't see myself making much headway from here...