SHERLOCK: Work around bad animation frames (bug #13101)
Apparently there are bad frames at the end of the first animation when
picking up the matchbook from the lab table. Some of these frames are
bad enough to crash ScummVM now that surface dimensions are signed. But
even before that, one of them would seriously slow down ScummVM, and
probably eat a lot of memory in the process.

This hard-codes an upper limit to the number of frames for the 59MATCH4
animation. Note that it will still include two of the bad frames to
avoid a glitch at the end of the animation, but these aren't
catastrophically bad frames. They're just weird.

It appears the original engine did not pre-decode the frames, so there
it was never a problem.